test(auth): divide tests by modules
